:lipstick: 一覧表示画面のメインテキストボックスを作成
